feat ui Relocate Customization drawer to Profile menu (#461)
This change moves the theme customization feature. The floating settings button on the right side of the screen has been removed. This fixes a bug where the button would incorrectly appear on the login and register pages.

The customization drawer state has been integrated into the global ConfigContext. The Customization component is now rendered inside the main AppLayout which ensures it is only available to logged in users.

A new Appearance button has been added to the ProfileSection menu. Clicking this button now opens the customization drawer. This provides a cleaner UI and places the theme settings in a more logical and context aware location.

I'm ready for the next enhancement.